Algorithms II
Let us consider a long, straight country road with n houses scattered sparsely along it. (We can picture the road as a long line segment, with a western endpoint and an eastern endpoint, where the western point is at position 0, the eastern point is at position L > 0, and the n houses are at positions x1, x2, …, xn, respectively, where 0 <= x1 < x2 < … < xn <= L.)
You want to place cell phone base stations at certain points along the road, so that every house is within (that is, less than or equal to) k miles of at least one of the base stations. Design an efficient algorithm that achieves this goal, using as few base stations as possible. Prove the correctness of your algorithm, and analyze its time complexity.