Algorithms3
We have three containers whose sizes are A pints, B pints and C pints, respectively, where A, B, C are all positive integers. In the beginning, the A-pint container has a pints of water, the B-pint container has b pints of water, and the C-pint container has c pints of water, where a, b, and c are non-negative integers. (For example, we might have A=10, B=7, C=4 and a=0, b=6, c=4.) We are allowed one type of operation: pouring the contents of one container into another, stopping only when the source container is empty or the destination container is full. We want to know if there is a sequence of pourings that, in the end, leaves exactly k pints of water in any of the three containers. (So the answer we seek is either YES or NO.)
Your task: 
1.) Model this problem as a graph problem: give a precise definition of the graph involved, and state the specific question about this graph that needs to be answered. 
2.) Design an efficient algorithm to solve the above problem.